+++ packages/expo-file-system/src/next/index.ts
@@ -1,0 +1,1 @@
+export { Directory, File, write } from './FileSystem';